fixed issue: dwarf unable to capture troll
self-reporting dwarf captures to backend.  trolls do not self report
captures because this game will operate on the assumption of compulsory
capturing.  without explicitly listing troll caps, the backend will auto
capture it in the notation.

this is only relevant because trolls always land in empty spaces but
dwarves land in occupied spaces for capping.